+++ libs/base_std/src/data/codec/hash/reexports.rs
@@ -3,9 +3,10 @@
 //!
 //
 
-use crate::_reexport;
+use crate::{_reexport, TAG_HASH};
 
 _reexport! { rust: std::hash,
+    tag: TAG_HASH!(),
     doc: "The default state for [`HashMapStd`][crate::HashMapStd].",
     RandomState
 }