+++ src/code/error/data.rs
@@ -23,28 +23,31 @@
 //   - MismatchedBounds: DataOverflow, IndexOutOfBounds, MismatchedIndices, MismatchedCapacity
 //   - PartialSpace:     NotEnoughSpace, PartiallyAdded
 
-use crate::{Interval, Mismatch, define_error};
+use crate::{Interval, Mismatch, TAG_DATA, define_error};
 
 /* individual errors */
 
 define_error! { individual: pub struct DataOverflow(pub Option<usize>);
-    // +tag: TAG_DATA!(), // MAYBE
+    +tag: TAG_DATA!(),
     DOC_DATA_OVERFLOW = "The value has surpassed the bounds of the representable data space.",
     self+f => if let Some(v) = self.0 {
         write!(f, "The value {v} has surpassed the bounds of the representable data space.")
     } else { write!(f, "The value has surpassed the bounds of the representable data space.") }
 }
 define_error! { individual: pub struct ElementNotFound;
+    +tag: TAG_DATA!(),
     DOC_ELEMENT_NOT_FOUND = "The requested element has not been found.",
     self+f => write!(f, "The requested element has not been found."),
 }
 define_error! { individual: pub struct IndexOutOfBounds(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_INDEX_OUT_OF_BOUNDS = "The given index is out of bounds.\n\n
 Optionally contains the given index.",
     self+f => if let Some(i) = self.0 { write!(f, "The given index {i} is out of bounds.")
     } else { write!(f, "The given index is out of bounds.") }
 }
 define_error! { individual: pub struct InvalidAxisLength(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_INVALID_AXIS_LENGTH = "The given axis has an invalid length.\n\n
 Optionally contains the given axis number.",
     self+f => if let Some(n) = self.0 {
@@ -52,10 +55,12 @@ Optionally contains the given axis number.",
     } else { write!(f, "One ore more axis have 0 length, which is not allowed.") }
 }
 define_error! { individual: pub struct KeyAlreadyExists;
+    +tag: TAG_DATA!(),
     DOC_KEY_ALREADY_EXISTS = "The key already exists.",
     self+f => write!(f, "The key already exists.")
 }
 define_error! { individual: pub struct MismatchedCapacity(pub Mismatch<Interval<usize>, usize>);
+    +tag: TAG_DATA!(),
     DOC_MISMATCHED_CAPACITY = "The given capacity did not match the required constraints.",
     self+f => write!(f, "Mismatched capacity: {:?}.", self.0)
 }
@@ -72,29 +77,35 @@ impl MismatchedCapacity {
     }
 }
 define_error! { individual: pub struct MismatchedDimensions(pub Mismatch<usize, usize>);
+    +tag: TAG_DATA!(),
     DOC_MISMATCHED_DIMENSIONS = "The dimensions given did not match the elements provided.",
     self+f => write!(f, "Mismatched dimensions: {:?}.", self.0)
 }
 define_error! { individual: pub struct MismatchedIndices;
+    +tag: TAG_DATA!(),
     DOC_MISMATCHED_INDICES = "The given indices does not match the expected order.",
     self+f => write!(f, "The given indices does not match the expected order.")
 }
 define_error! { individual: pub struct NodeEmpty(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_NODE_EMPTY = "The node is empty.",
     self+f => if let Some(n) = self.0 { write!(f, "The given node `{n}` is empty.")
     } else { write!(f, "The node is empty.") }
 }
 define_error! { individual: pub struct NodeLinkNotSet(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_NODE_LINK_NOT_SET = "The link is not set.",
     self+f => if let Some(n) = self.0 { write!(f, "The given node link `{n}` is not set.")
     } else { write!(f, "The node link is not set.") }
 }
 define_error! { individual: pub struct NodeLinkNotUnique(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_NODE_LINK_NOT_UNIQUE = "The link is not unique.",
     self+f => if let Some(n) = self.0 { write!(f, "The given node link `{n}` is not unique.")
     } else { write!(f, "The node link is not unique.") }
 }
 define_error! { individual: pub struct NotEnoughElements(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_NOT_ENOUGH_ELEMENTS = "There are not enough elements for the operation.\n\n
 Optionally contains the minimum number of elements needed.",
     self+f => if let Some(n) = self.0 {
@@ -102,6 +113,7 @@ Optionally contains the minimum number of elements needed.",
     } else { write!(f, "Not enough elements.") }
 }
 define_error! { individual: pub struct NotEnoughSpace(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_NOT_ENOUGH_SPACE = "There is not enough free space for the operation.\n\n
 Optionally contains the number of free spaces needed.",
     self+f => if let Some(n) = self.0 {
@@ -109,6 +121,7 @@ Optionally contains the number of free spaces needed.",
     } else { write!(f, "Not enough space.") }
 }
 define_error! { individual: pub struct PartiallyAdded(pub Option<usize>);
+    +tag: TAG_DATA!(),
     DOC_PARTIALLY_ADDED = "The operation could only add a subset of the elements.\n\n
 Optionally contains the number of elements added.",
     self+f => if let Some(n) = self.0 { write!(f, "Only `{n}` elements could be added.")